Change the logic of driving, edit classes Car, CarMechanic and interfaces ICarMechanics, IDriveable correspondly.

+++ CarRental/Models/Car/ICarMechanics.cs
@@ -17,6 +17,6 @@ public interface ICarMechanics
     public ComponentStatus CheckComponent(IComponent component);
     public bool TryFixComponent(IComponent component);
 
-    internal int CheckFuel(Car car);
+    internal float CheckFuel(Car car);
     internal void Refill(Car car, int amountFuel);
 }
